requirements.txt
# Core DL stack
torch
torchvision
torchaudio

# Utils
numpy
matplotlib
scikit-learn
tqdm
pyyaml
pillow
